+++ src/services/scroll-resolver.ts
@@ -1,6 +1,6 @@
-import { IPositionStats, IScrollerConfig } from '../models';
+import { IPositionStats, IScrollerConfig, IScrollState } from '../models';
 
-export function shouldScroll(
+export function shouldFireScrollEvent(
   container: IPositionStats,
   config: IScrollerConfig,
   scrollingDown: boolean
@@ -9,14 +9,15 @@ export function shouldScroll(
   let remaining: number;
   let containerBreakpoint: number;
   if (scrollingDown) {
-    remaining = container.totalToScroll - container.scrolledUntilNow;
-    containerBreakpoint = container.height * distance.down + 1;
+    remaining = (container.totalToScroll - container.scrolledUntilNow) / container.totalToScroll;
+    containerBreakpoint = distance.down / 10;
   } else {
-    remaining = container.scrolledUntilNow;
-    containerBreakpoint = container.height * distance.up + 1;
+    remaining = container.scrolledUntilNow / container.totalToScroll;
+    containerBreakpoint = distance.up / 10;
   }
-  const shouldScroll: boolean = remaining <= containerBreakpoint;
-  return shouldScroll;
+
+  const shouldFireEvent: boolean = remaining <= containerBreakpoint;
+  return shouldFireEvent;
 }
 
 export function isScrollingDownwards(
@@ -33,11 +34,33 @@ export function getScrollStats(
 ) {
   const isScrollingDown = isScrollingDownwards(lastScrollPosition, container);
   return {
-    shouldScroll: shouldScroll(container, config, isScrollingDown),
+    shouldFireScrollEvent: shouldFireScrollEvent(container, config, isScrollingDown),
     isScrollingDown
   };
 }
 
-export function updateScrollPosition(position: number, lastPositionState) {
-  return (lastPositionState.last = position);
+export function updateScrollPosition(position: number, scrollState: IScrollState) {
+  return (scrollState.lastScrollPosition = position);
+}
+
+export function updateTotalToScroll(totalToScroll: number, scrollState: IScrollState) {
+  scrollState.lastTotalToScroll = scrollState.totalToScroll;
+  scrollState.totalToScroll = totalToScroll;
+}
+
+export function isSameTotalToScroll(scrollState) {
+  return scrollState.totalToScroll === scrollState.lastTotalToScroll;
+}
+
+export function updateTriggeredFlag(scrollState, triggered: boolean) {
+  scrollState.isTriggeredTotal = triggered;
+}
+
+export function updateScrollState(scrollState: IScrollState, scrolledUntilNow: number, totalToScroll: number) {
+  updateScrollPosition(scrolledUntilNow, scrollState);
+  updateTotalToScroll(totalToScroll, scrollState);
+  const isSameTotal = isSameTotalToScroll(scrollState);
+  if (!isSameTotal) {
+    updateTriggeredFlag(scrollState, false);
+  }
 }